v2/session/web/requirements.txt
Flask-And-Redis
flask-cors
flask-restful